finish porting everything to the new command system
